Fix clamping issues
Previously, when the vector that holds the list of window ids that a
user has passed through(forwards or backwards) is more than a set value,
it gets truncated to a specific length. However, this truncation occurs
at the end of the vector, thus only removing the most recent window ids,
when it should be removing the oldest ones.